Update statistikk schemas, improve date handling, and add new field
Also cleaned up the mapper code so that it's hopefully easier to navigate
