+++ service/src/main/resources/statistikk/sak_schema.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "http://json-schema.org/draft-04/schema#",
+  "$schema": "http://json-schema.org/draft-07/schema#",
   "title": "sak",
   "description": "Det overordnede behandlingsforløp knyttet til én bruker, referert til det samme rettighetsforholdet. Saken henviser til den overordnede mappen av behandlinger.",
   "required": [
